+++ my-app/src/components/PokemonCard/index.js
@@ -1,27 +1,28 @@
 import { useState } from 'react';
 import s from './style.module.css';
 import cardBackSide from './assets/card-back-side.jpg';
+import cn from 'classnames';
 
 const PokemonCard = ({name, img, id, type, values}) => {
 
 const [isActive, setActive] = useState(false);
 
 const handleClick = () => {
-    setActive(true);
+    setActive(!isActive);
 }
 
 return (
 <>
 <div className={s.root} onClick={handleClick}>
-      <div className={`${s.pokemonCard} ${isActive? s.active : ''}`}>
+      <div className={cn(s.pokemonCard, {[s.active]: isActive})}>
         <div className={s.cardFront}>
-            <div className={`${s.wrap} ${s.front}`}>
-                <div className={`${s.pokemon} ${s[type]}`}>
+            <div className={cn(s.wrap, s.front)}>
+                <div className={cn(s.pokemon, s.type)}>
                     <div className={s.values}>
-                        <div className={`${s.count} ${s.top}`}>{values.top}</div>
-                        <div className={`${s.count} ${s.right}`}>{values.right}</div>
-                        <div className={`${s.count} ${s.bottom}`}>{values.bottom}</div>
-                        <div className={`${s.count} ${s.left}`}>{values.left}</div>
+                        <div className={cn(s.count, s.top)}>{values.top}</div>
+                        <div className={cn(s.count, s.right)}>{values.right}</div>
+                        <div className={cn(s.count, s.bottom)}>{values.bottom}</div>
+                        <div className={cn(s.count, s.left)}>{values.left}</div>
                     </div>
                     <div className={s.imgContainer}>
                         <img src={img} alt={name} />
@@ -38,7 +39,7 @@ return (
         </div>
 
         <div className={s.cardBack}>
-            <div className={`${s.wrap} ${s.back}`}>
+            <div className={cn(s.wrap, s.back)}>
                 <img src={cardBackSide} alt="Сard Backed" />
             </div>
         </div>
